Restructured codebase, removed unused files
