--- frontend/my-frontend/src/components/portfolio.js
@@ -1,24 +1,0 @@
-const projects = [
-  {
-    name: "GoSports",
-    subName: "Full Stack Django Web App",
-    projectLogo: "../static/pics/blue_go_sports.png",
-    techStackImgs: [
-      "../static/pics/python-5.svg",
-      "../static/pics/django.jpg",
-      "../static/pics/html5-2.svg",
-      "../static/pics/bootstrap.svg",
-    ],
-    projectDesc:
-      "Web app that provides you everything NBA. Created with a Django backend and a bootstrap frontend, you can view real time live data",
-    featuresList: [
-      "view live scores of games going on today",
-      "Look at your favorite teams roster, as well as each players indivual stats that season",
-      "See current NBA standings",
-      "View each active nba players career, also see common info about that player",
-      "Search for your favorite player or team",
-      "Create an account",
-    ],
-    gitHubLink: "https://github.com/sjdefran1/GoSports",
-  },
-];
